+++ PicesLibraries/PicesLibrary/svm2.cpp
@@ -4242,7 +4242,7 @@ void  SVM289_MFS::svm_save_model_XML (ostream&          o,
                                   const svm_model&  model
                                  )
 {
-  int32  origPrecision = o.precision ();
+  int32  origPrecision = (int32)o.precision ();
 
   int32 nr_class = model.nr_class;
   int32 l = model.l;
@@ -5015,7 +5015,7 @@ void  SVM289_MFS::svm_model::Write (ostream& o)
     const  FeatureVector&  p = SV[i];
     o << "SupportVector" << "\t" << p.ImageFileName ();
 
-    int32  origPrec = o.precision ();
+    int32  origPrec = (int32)o.precision ();
     o.precision (16);
     for  (int32 j = 0;  j < nr_class - 1;  j++)
     {
